fix: v0.2.4 — Windows HANDLE 类型修复
windows-sys 0.59 的 HANDLE 是 *mut c_void 而非 isize。
移除所有 as_raw_handle() as isize 转型，RawHandle 本身已是正确类型。

- Remove `as isize` casts in win/mod.rs, win/pseudocon.rs, lib.rs
- Wrap raw handle in SendHandle for thread safety in Future impl

+++ src/lib.rs
@@ -295,7 +295,7 @@ impl ChildKiller for ProcessSignaller {
             use std::os::windows::io::AsRawHandle;
             unsafe {
                 if windows_sys::Win32::System::Threading::TerminateProcess(
-                    handle.as_raw_handle() as isize,
+                    handle.as_raw_handle(),
                     127,
                 ) == 0
                 {
